Note taking application
This application is an exectuable JAR file. It opens up a gui and presents you with the option to create a note. It asks you for a title for your note and then asks you to to fill in the contents. once you click save a note is created and saved(made for my personal directory)  its not finished but a good start